requirements.txt
paddleocr>=2.7.0
paddlepaddle>=2.5.0
numpy>=1.23.0
Pillow>=9.5.0
streamlit>=1.28.0
